+++ src/main.h
@@ -7,6 +7,6 @@
 #define BUFFER_SIZE 1048576
 #define NEW_ENCODE 2000
 #define NEW_DECODE 3000
-int main(uint64_t argc, char *argv[]);
-uint64_t parse_commandline_args(uint64_t argc, char *argv[]);
+int main(int argc, char *argv[]);
+int parse_commandline_args(int argc, char *argv[]);
 void print_bad_input_message(char *message);
